+++ refinedstorage2-query-parser/src/main/java/com/refinedmods/refinedstorage2/query/lexer/LexerTokenMappings.java
@@ -8,6 +8,7 @@ public class LexerTokenMappings {
             .addMapping(new LexerTokenMapping("!", TokenType.UNARY_OP))
             .addMapping(new LexerTokenMapping("@", TokenType.UNARY_OP))
             .addMapping(new LexerTokenMapping("$", TokenType.UNARY_OP))
+            .addMapping(new LexerTokenMapping("#", TokenType.UNARY_OP))
             .addMapping(new LexerTokenMapping(">", TokenType.UNARY_OP))
             .addMapping(new LexerTokenMapping(">=", TokenType.UNARY_OP))
             .addMapping(new LexerTokenMapping("<", TokenType.UNARY_OP))